requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.5
google-cloud-vision==3.4.4
pydantic==2.5.0
python-dotenv==1.0.0
pdf2image==1.16.3
Pillow==10.1.0
google-generativeai==0.3.2
openai
langchain-openai 